requirements.txt
numpy
glfw
Pillow
PyOpenGL
PyOpenGL_accelerate